+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+0.5.4
+-----
+- Added marker option to loads, so that load markers are optional
+- OpenID client will now extract tokens from cookies as well as the header.
+
 0.5.3
 -----
 - Added utility function integrate-ident!